seperated inputs and outputs
