+++ src/strain_relief/io/__init__.py
@@ -1,10 +1,8 @@
-from ._io import load_parquet, save_parquet, to_mols_dict
-from .utils_mol_format import ase_to_rdkit, rdkit_to_ase
+from ._input import load_parquet, to_mols_dict
+from ._output import save_parquet
 
 __all__ = [
     "load_parquet",
     "save_parquet",
     "to_mols_dict",
-    "ase_to_rdkit",
-    "rdkit_to_ase",
 ]
